+++ lib/blackmagicdesign.dart
@@ -1,3 +1,4 @@
+/// Utilities for communicating with Blackmagic Design broadcast devices.
 library blackmagicdesign;
 
 export 'atem/atem.dart';